backend/requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Machine Learning and NLP
torch>=2.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
numpy==1.24.4
pandas==2.1.4

# Document processing and PDF generation
weasyprint==60.2
markdown==3.5.1
pypdf2==3.0.1
python-docx==1.1.0

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0

# HTTP requests and API integration
requests==2.31.0
aiohttp==3.9.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# Utilities
tqdm==4.66.1
regex==2023.12.25

# Database and ORM
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Google OAuth and Drive API
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.88.0

# Session management and security
Flask-Session==0.5.0
cryptography==41.0.7
redis==4.5.5

# Additional utilities for auth
PyJWT==2.8.0
bcrypt==4.0.1

# Security and validation
bleach==6.1.0
jsonschema==4.20.0

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.0
flake8==6.1.0